+++ src/idtrackerai_GUI_tools/GUI_main_base.py
@@ -100,7 +100,16 @@ class GUIBase(QMainWindow):
 
     def center_window(self):
         w, h = 1000, 800
-        cp = QGuiApplication.screenAt(self.cursor().pos()).availableGeometry().center()
+        try:
+            cp = (
+                QGuiApplication.screenAt(self.cursor().pos())
+                .availableGeometry()
+                .center()
+            )
+        except AttributeError:
+            # in Fedora QGuiApplication.screenAt(self.cursor().pos()) is None
+            cp = QGuiApplication.primaryScreen().availableGeometry().center()
+
         self.setGeometry(cp.x() - w // 2, cp.y() - h // 2, w, h)
 
     def change_theme(self, dark_theme: bool):